Fix some warnings

+++ src/Detector/tensorrt_yolo/YoloONNXv8_bb.hpp
@@ -34,7 +34,7 @@ protected:
 
 		size_t ncInd = 1;
 		size_t lenInd = 2;
-		int nc = m_outpuDims[0].d[ncInd] - 4;
+		int nc = static_cast<int>(m_outpuDims[0].d[ncInd] - 4);
 		int dimensions = nc + 4;
 		size_t len = static_cast<size_t>(m_outpuDims[0].d[lenInd]) / m_params.explicitBatchSize;
 		//auto Volume = [](const nvinfer1::Dims& d)
